Add files via upload
Was able to map the phase curve such that the x-axis represents time-series data, not just as a function of angle. This was done by using the vis viva equation to determine the instantaneous velocity of the orbit at every point, and using an elliptic integral to determine the ellipse circumference. The theta and time arrays are indexically 1:1, such that a given theta maps to the corresponding time based on how much arc length the planet has subtended from the vis viva equation. This has the distinct advantage of being able to switch between the theta array and the time series array for the x-axis in plotting. In essence, the only thing that has occurred is that the spacing of the theta array has been mapped to a function which transforms it into the time domain. Consequently, it shifts the location of periastron and of the transit off 0, which is expected; however, periastron is consistent across all functions, and I have made it such that the periastron index (the float between -1 and 1 where periastron is in the phase curve) is produced for both the angle and the time series array, and are labeled accordingly in the output.